hdf5 reader: add docstring + split Resources.py
